Refactoring the stress test orchestrations into a separate assemly so that service fabric implementation can reference those and stress test executable can reference the service fabric.
