+++ Test/DurableTask.Test.Orchestrations/Stress/TestOrchestration.cs
@@ -11,7 +11,7 @@
 //  limitations under the License.
 //  ----------------------------------------------------------------------------------
 
-namespace DurableTask.Stress.Tests
+namespace DurableTask.Test.Orchestrations.Stress
 {
     using System;
     using System.Collections.Generic;
@@ -19,25 +19,34 @@ namespace DurableTask.Stress.Tests
     using System.Threading.Tasks;
     using DurableTask;
 
-    public class DriverOrchestration : TaskOrchestration<int, DriverOrchestrationData>
+    public class TestOrchestration : TaskOrchestration<int, TestOrchestrationData>
     {
-        public override async Task<int> RunTask(OrchestrationContext context, DriverOrchestrationData data)
+        public override async Task<int> RunTask(OrchestrationContext context, TestOrchestrationData data)
         {
             int result = 0;
             List<Task<int>> results = new List<Task<int>>();
             int i = 0;
+            int j = 0;
             for (; i < data.NumberOfParallelTasks; i++)
             {
-                results.Add(context.CreateSubOrchestrationInstance<int>(typeof(TestOrchestration), data.SubOrchestrationData));
+                results.Add(context.ScheduleTask<int>(typeof(TestTask), new TestTaskData
+                {
+                    TaskId = "ParallelTask: " + i.ToString(),
+                    MaxDelayInMinutes = data.MaxDelayInMinutes,
+                }));
             }
 
             int[] counters = await Task.WhenAll(results.ToArray());
             result = counters.Max();
 
-            if (data.NumberOfIteration > 1)
+            for (; j < data.NumberOfSerialTasks; j++)
             {
-                data.NumberOfIteration--;
-                context.ContinueAsNew(data);
+                int c = await context.ScheduleTask<int>(typeof(TestTask), new TestTaskData
+                {
+                    TaskId = "SerialTask" + (i + j).ToString(),
+                    MaxDelayInMinutes = data.MaxDelayInMinutes,
+                });
+                result = Math.Max(result, c);
             }
 
             return result;
