+++ Test/DurableTask.Test.Orchestrations/Stress/TestTask.cs
@@ -11,12 +11,11 @@
 //  limitations under the License.
 //  ----------------------------------------------------------------------------------
 
-namespace DurableTask.Stress.Tests
+namespace DurableTask.Test.Orchestrations.Stress
 {
     using System;
     using System.Threading;
     using System.Threading.Tasks;
-    using DurableTask;
 
     public sealed class TestTask : AsyncTaskActivity<TestTaskData, int>
     {
