+++ test/DurableTask.Stress.Tests/DurableTask.Stress.Tests.csproj
@@ -35,6 +35,10 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <ProjectReference Include="..\DurableTask.Test.Orchestrations\DurableTask.Test.Orchestrations.csproj">
+      <Project>{2bc2a5ad-4200-41b9-a034-332bb1d7187b}</Project>
+      <Name>DurableTask.Test.Orchestrations</Name>
+    </ProjectReference>
     <Reference Include="CommandLine">
       <HintPath>$(SolutionDir)\packages\CommandLineParser.1.9.71\lib\net45\CommandLine.dll</HintPath>
     </Reference>
@@ -93,17 +97,11 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DriverOrchestration.cs" />
-    <Compile Include="DriverOrchestrationData.cs" />
     <Compile Include="Options.cs" />
     <Compile Include="OrchestrationConsoleTraceListener.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestOrchestration.cs" />
-    <Compile Include="TestOrchestrationData.cs" />
-    <Compile Include="TestTask.cs" />
     <Compile Include="OrchestrationFileTraceListener.cs" />
-    <Compile Include="TestTaskData.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config">
